% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcmap-methods.R
\name{plot.spcmap}
\alias{plot.spcmap}
\title{Plot a spcmap object}
\usage{
plot.spcmap(x, method = c("ggplot2"))
}
\arguments{
\item{x}{An object of class \code{spcmap}.}

\item{method}{The plotting method. Can be \code{"ggplot2"} for \code{ggplot2}-based
plots.}
}
\description{
Plot a spcmap object
}
