% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcmap.R
\name{spcmap}
\alias{spcmap}
\title{Compare sample coordinates with NOAA Storm Prediction Center risk maps}
\usage{
spcmap(data, day = 1, type = c("cat", "torn", "hail", "wind"))
}
\arguments{
\item{data}{A \code{data.frame} with information on sample coordinates. The
object must have the columns \code{lat} and \code{long} and may have any
number of additional columns.}

\item{day}{The \code{integer} day of the outlook. Must be at least \code{1}
but no more that \code{3}.}

\item{type}{The type of convective outlook data. \code{"cat"} is categorical,
\code{"torn"} is tornado risk, \code{"hail"} is hail risk, and \code{"wind"}
is wind risk. Note that only \code{"cat"} can be accepted if \code{day} is
greater than \code{1}.}
}
\description{
Compare sample coordinates with NOAA Storm Prediction Center risk maps
}
