% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcmap-methods.R
\name{summary.spcmap}
\alias{summary.spcmap}
\alias{print.spcmap}
\title{Summarize a spcmap object}
\usage{
summary.spcmap(x)

\method{print}{spcmap}(x)
}
\arguments{
\item{x}{An object of class \code{spcmap}.}
}
\description{
Summarize a spcmap object

Print a spcmap object
}
